+++ packages/api-contract/src/index.ts
@@ -25,3 +25,4 @@ export type BalanceSnapshotList = components['schemas']['BalanceSnapshotList'];
 export type RecordBalanceSnapshotRequest = components['schemas']['RecordBalanceSnapshotRequest'];
 export type BalanceReminder = components['schemas']['BalanceReminder'];
 export type BalanceReminderSetting = components['schemas']['BalanceReminderSetting'];
+export type MonthlyDashboard = components['schemas']['MonthlyDashboard'];